refactor: 라우팅 설정 및 폴더 구조 재배치
