--- frontend/src/App.tsx
@@ -1,11 +1,0 @@
-import PlaylistCard from "./components/PlaylistCard/PlaylistCard";
-
-function App() {
-  return (
-    <div className="min-h-screen flex bg-tomato items-center justify-center">
-      <PlaylistCard />
-    </div>
-  );
-}
-
-export default App;
