+++ frontend/src/routes/path.ts
@@ -1,0 +1,3 @@
+export const PATHS = {
+  ROOM: "/room",
+} as const;
