+++ frontend/src/types/todo.ts
@@ -1,0 +1,11 @@
+export interface TodoItem {
+  id: number;
+  label: string;
+  done: boolean;
+}
+
+export interface TodoGroup {
+  id: string;
+  title: string;
+  items: TodoItem[];
+}
